refactor: move watch functions to dedicated file (#423)
This adds a clear separation between the actions API client and the
actions watch functions.

(cherry picked from commit 575c24c57e5a38c48ca747864753b43cd7a8261b)
